controller level restrictions on editing and deleting articles

+++ app/controllers/articles_controller.rb
@@ -27,10 +27,17 @@ class ArticlesController < ApplicationController
   end
   
   def edit
-
+    unless @article.user == current_user
+      flash[:alert] = "You can only edit your own article"
+      redirect_to root_path
+    end
   end
 
   def update
+    unless @article.user == current_user
+      flash[:danger] = "unless @article.user == current_user"
+      redirect_to root_path
+    else
       if @article.update(article_params)
         flash[:success] = "Article has been updated"
         redirect_to @article 
@@ -38,12 +45,18 @@ class ArticlesController < ApplicationController
         flash.now[:danger] = "Article has not been updated"
         render :edit
       end
+    end
   end
 
   def destroy
-    if @article.destroy
-      flash[:success] = "Article has been deleted."
-      redirect_to articles_path 
+    unless @article.user == current_user
+      flash[:danger] = "unless @article.user == current_user"
+      redirect_to root_path
+    else
+      if @article.destroy
+        flash[:success] = "Article has been deleted."
+        redirect_to articles_path 
+      end
     end
   end
